Moving lang2001 to the new lang_2001 package for consistent naming with lang_brainfuck
